// packages/extras/DataTable/src/components/DataTable/type.ts
type CellPropsType = TableCellPropsType & Record<string, any>;

export type DataTableSortType = 'numerical' | 'alphabetical';
// ...
export type DataTableColumnType<
    D extends DataTableDataType = DataTableDataType,
> = {
    /** Column identifier */
    id: string;
    /** Column title */
    title?: ReactChild;
    key?: string;

    /** Indicates whether this columnLayout is sortable or not */
    sortable?: boolean;
    /** Type of sorting for this column */
    sortType?: DataTableSortType;
    /** Properties to be forwarded to every header cell */
    headCellProps?: CellPropsType;
    /** Properties to be forwarded to every data cell */
    dataCellProps?: CellPropsType;
    /** Properties to be forwarded to every cell */
    cellProps?: TableCellPropsType & Record<string, any>;
    /** If set to true, the columnLayout acts as a controller for the expandable section */
    expandableSectionController?: boolean;
    /** Same as dataCellProps, but can be set individually for each table row */
    getDataCellProps?: (
        value: any,
        item: D,
        column: DataTableColumnType<D>,
    ) => CellPropsType;
    /** Renders the content of a cell for each row individually */
    formatDataCellContent?: (
        value: any,
        item: D,
        column: DataTableColumnType<D>,
    ) => ReactChild | ReactChild[];

    ellipsis?: boolean;
    width?: ScalarType;
    contentAlignY?: 'top' | 'center' | 'bottom';
};

export type DataTableDataType = {
    /** Row identifier */
    id: ScalarType;
    key?: ScalarType;
} & Record<string, any>;